+++ changelog.md
@@ -2,3 +2,5 @@ Appota Game SDK ver 4.0.7b7
 Update :
 + fixed sdk crash on iOS6
 + fixed payment via techcombank
+CLUpdate02:
++ fix error code 401 : unauthorized